Delete lecture router and added folders
