+++ src/routers/Questions & Responses/GetResponsesRouter.js
@@ -3,11 +3,11 @@ import type {
   SingleResponse,
   MultipleResponse,
   APIAnswer
-} from './APITypes';
+} from '../APITypes';
 
-import AppDevEdgeRouter from '../utils/AppDevEdgeRouter';
-import constants from '../utils/constants';
-import ResponsesRepo from '../repos/ResponsesRepo';
+import AppDevEdgeRouter from '../../utils/AppDevEdgeRouter';
+import constants from '../../utils/constants';
+import ResponsesRepo from '../../repos/ResponsesRepo';
 import { Request } from 'express';
 
 class GetResponsesRouter extends AppDevEdgeRouter<APIAnswer> {
